# internal/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Admin API of the goose proxy-pool engine: create and manage inbounds (http/socks5 listeners), outbounds (proxy plugins), pools and chains, and inspect request metrics.",
        "title": "goose admin API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/api/chains": {
            "get": {
                "description": "Returns all configured proxy chains (ordered lists of pool layers).\nReturns the chain with the given id.\nCreates (or replaces) a proxy chain: an ordered list of layers, each referencing a pool (which carries its own filters + selector). Layer 0 is the first hop; each subsequent layer is dialed through the previous.\nRemoves the chain with the given id. Inbounds whose policy still references it will fail to resolve until reconfigured.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "application/json",
                    "application/json"
                ],
                "tags": [
                    "chains",
                    "chains",
                    "chains",
                    "chains"
                ],
                "summary": "Delete a chain",
                "parameters": [
                    {
                        "description": "Chain to create",
                        "name": "chain",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/config.ChainSpec"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/config.ChainSpec"
                        }
                    },
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/config.ChainSpec"
                        }
                    },
                    "204": {
                        "description": "No content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.errResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.errResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.errResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Returns all configured proxy chains (ordered lists of pool layers).\nReturns the chain with the given id.\nCreates (or replaces) a proxy chain: an ordered list of layers, each referencing a pool (which carries its own filters + selector). Layer 0 is the first hop; each subsequent layer is dialed through the previous.\nRemoves the chain with the given id. Inbounds whose policy still references it will fail to resolve until reconfigured.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "application/json",
                    "application/json"
                ],
                "tags": [
                    "chains",
                    "chains",
                    "chains",
                    "chains"
                ],
                "summary": "Delete a chain",
                "parameters": [
                    {
                        "description": "Chain to create",
                        "name": "chain",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/config.ChainSpec"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/config.ChainSpec"
                        }
                    },
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/config.ChainSpec"
                        }
                    },
                    "204": {
                        "description": "No content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.errResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.errResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.errResponse"
                        }
                    }
                }
            }
        },
...
        "/api/engine": {
            "get": {
                "description": "Returns the engine-level configuration: network stack selection, admin API settings, metrics DB path and offline geo database settings.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "engine"
                ],
                "summary": "Get engine configuration",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/config.Engine"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.errResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.errResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error is a short, human-readable reason for the rejection.",
                    "type": "string",
                    "example": "missing id"
                }
            }
        },
        "config.APIConfig": {
            "type": "object",
            "properties": {
                "listen": {
                    "type": "string"
                },
                "token": {
                    "description": "Token is the bearer token for admin auth; empty disables auth.",
                    "type": "string"
                }
            }
        },
        "config.ChainSpec": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "layers": {
                    "description": "PoolIDs, in order",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "config.Engine": {
            "type": "object",
            "properties": {
                "api": {
                    "description": "API is the admin server config.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/config.APIConfig"
                        }
                    ]
                },
                "db": {
                    "description": "DB is the metrics database path.",
                    "type": "string"
                },
                "geo": {
                    "description": "Geo is the offline IP database config.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/config.GeoConfig"
                        }
                    ]
                },
                "stack": {
                    "description": "Stack selects the network stack: \"system\" or \"gvisor\".",
                    "type": "string"
                }
            }
        },
...
        "config.GeoConfig": {
            "type": "object",
            "properties": {
                "path": {
                    "description": "Path is the database file path.",
                    "type": "string"
                },
                "type": {
                    "description": "Type is \"mmdb\", \"qqwry\", \"zxinc\" or \"\" (disabled).",
                    "type": "string"
                }
            }
        },
...
    },
...
}